# lib/api-zod/src/generated/types/brand.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AffiliateDeals API
 * OpenAPI spec version: 0.1.0
 */

export interface Brand {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  offerCount?: number;
  isActive: boolean;
}