# lib/api-zod/src/generated/types/brandUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AffiliateDeals API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  description?: string;
  logoUrl?: string;
  websiteUrl?: string;
  isActive?: boolean;
}